feat: Bootstrap new bruno project
Bruno is a REST client that is similar to the better known "Postman",
but simpler, cleaner, open source. https://www.usebruno.com/

This project holds a collection. The collection allows us to test the
issuance and importing of Verifiable credentials, of type "OpenBadges".

The collection does not adhere to the specs yet. But the plan is to make
it do so. It is work in progress.

For example it hardcodes some logic that should be dynamic, according to
the spec. The plan is to remove these hardcoded paths and make them
follow the spec.

It also follows some hardcoded paths, that should, according to the
spec, be determined from other request like .well-known documents. The
agents used to test not all offer these according to spec - yet.

It has some business logic to sign and verify credentials,  but the
actual business logic goes beyond what is easy to achieve in bruno.
Ideally we would make libraries for this and call them instead, in
bruno. For now we keep things simple and don't adhere to the specs,
instead often have mock- logic.